:bomb: before tests
