+++ .dialyzer/ignore.exs
@@ -1,1 +1,3 @@
-[]
+[
+
+]
